packages/apps/pyproject.toml
[project]
name = "microsoft-teams-apps"
version = "2.0.0a3"
description = "The app package for a Microsoft Teams agent"
authors = [{ name = "Microsoft", email = "TeamsAISDKFeedback@microsoft.com" }]
readme = "README.md"
requires-python = ">=3.12"
repository = "https://github.com/microsoft/teams.py"
keywords = ["microsoft", "teams", "ai", "bot", "agents"]
license = "MIT"

dependencies = [
    "fastapi>=0.115.13",
    "microsoft-teams-api",
    "microsoft-teams-common",
    "uvicorn>=0.34.3",
    "cryptography>=3.4.0",
    "pyjwt[crypto]>=2.10.0",
    "dependency-injector>=4.48.1",
]

[project.optional-dependencies]
graph = [
    "microsoft-teams-graph",
]

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/microsoft", "scripts"]

[project.scripts]
generate-activity-handlers = "scripts.generate_handlers:generate_activity_handlers"

[tool.uv.sources]
microsoft-teams-common = { workspace = true }
microsoft-teams-api = { workspace = true }
microsoft-teams-graph = { workspace = true }
